backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu 
